#45: if no active option, set first one as active
the choosen option will not be located, only be highlighted. seems even locating the choosen option, system cannot make sure that the option which will be choosen is near the choosen one. so in this case, it is unnecessary now.

and for filter scenario, if there is an active option, and the option didn't filtered by filter text, then keep this option as active. if it is filtered, let first one as active